LPD-71983 Move to the correct package
